Extract usePeers and useCamera
