+++ packages/react-native-client/src/RNFishjamClient.types.ts
@@ -1,92 +1,13 @@
 import type { RTCStats } from './stats/types';
 import type { MicrophoneConfig } from './hooks/useMicrophone';
-
-export enum TrackType {
-  Audio = 'Audio',
-  Video = 'Video',
-}
-/**
- * Type describing Voice Activity Detection statuses.
- *
- * SPEECH - voice activity has been detected
- * SILENCE - lack of voice activity has been detected
- */
-export enum VadStatus {
-  Silence = 'silence',
-  Speech = 'speech',
-}
-
-/**
- * Type describing possible reasons of currently selected encoding.
- *
- * - OTHER - the exact reason couldn't be determined
- * - ENCODING_INACTIVE - previously selected encoding became inactive
- * - LOW_BANDWIDTH - there is no longer enough bandwidth to maintain previously selected encoding
- */
-export enum EncodingReason {
-  Other = 'other',
-  EncodingInactive = 'encodingInactive',
-  LowBandwidth = 'lowBandwidth',
-}
-
-export type Track<MetadataType extends Metadata> = {
-  id: string;
-  type: TrackType;
-  metadata: MetadataType;
-  vadStatus: VadStatus;
-  // Encoding that is currently received. Only present for remote tracks.
-  encoding: TrackEncoding | null;
-  // Information about simulcast, if null simulcast is not enabled
-  simulcastConfig: SimulcastConfig | null;
-  // The reason of currently selected encoding. Only present for remote tracks.
-  encodingReason: EncodingReason | null;
-};
+import type { Endpoint } from './hooks/usePeers';
+import type { CameraConfig, CaptureDevice } from './hooks/useCamera';
 // global
 export type Metadata = { [key: string]: any };
-
-export type Endpoint<
-  MetadataType extends Metadata,
-  VideoTrackMetadataType extends Metadata,
-  AudioTrackMetadataType extends Metadata,
-> = {
-  /**
-   *  id used to identify an endpoint
-   */
-  id: string;
-  /**
-   * used to indicate endpoint type.
-   */
-  type: string;
-  /**
-   * whether the endpoint is local or remote
-   */
-  isLocal: boolean;
-  /**
-   * a map `string -> any` containing endpoint metadata from the server
-   */
-  metadata: MetadataType;
-  /**
-   * a list of endpoints's video and audio tracks
-   */
-  tracks: Track<VideoTrackMetadataType | AudioTrackMetadataType>[];
-};
 
 export enum VideoLayout {
   FILL = 'FILL',
   FIT = 'FIT',
-}
-
-export enum VideoQuality {
-  QVGA_169 = 'QVGA169',
-  VGA_169 = 'VGA169',
-  QHD_169 = 'QHD169',
-  HD_169 = 'HD169',
-  FHD_169 = 'FHD169',
-  QVGA_43 = 'QVGA43',
-  VGA_43 = 'VGA43',
-  QHD_43 = 'QHD43',
-  HD_43 = 'HD43',
-  FHD_43 = 'FHD43',
 }
 
 export enum ScreencastQuality {
@@ -133,44 +54,6 @@ export type SimulcastBandwidthLimit = Record<TrackEncoding, BandwidthLimit>;
  */
 export type TrackBandwidthLimit = BandwidthLimit | SimulcastBandwidthLimit;
 
-export type CameraConfig<MetadataType extends Metadata> = {
-  /**
-   * resolution + aspect ratio of local video track, one of: `QVGA_169`, `VGA_169`, `QHD_169`, `HD_169`,
-   * `FHD_169`, `QVGA_43`, `VGA_43`, `QHD_43`, `HD_43`, `FHD_43`. Note that quality might be worse than
-   * specified due to device capabilities, internet connection etc.
-   * @default `VGA_169`
-   */
-  quality: VideoQuality;
-  /**
-   * whether to flip the dimensions of the video, that is whether to film in vertical orientation.
-   * @default `true`
-   */
-  flipVideo: boolean;
-  /**
-   * a map `string -> any` containing video track metadata to be sent to the server.
-   */
-  videoTrackMetadata: MetadataType;
-  /**
-   *  SimulcastConfig of a video track. By default simulcast is disabled.
-   */
-  simulcastConfig: SimulcastConfig;
-  /**
-   *  bandwidth limit of a video track. By default there is no bandwidth limit.
-   */
-  maxBandwidth: TrackBandwidthLimit;
-  /**
-   * whether the camera track is initially enabled, you can toggle it on/off later with toggleCamera method
-   * @default `true`
-   */
-  cameraEnabled: boolean;
-  /**
-   * id of the camera to start capture with. Get available cameras with `getCaptureDevices()`.
-   * You can switch the cameras later with `flipCamera`/`switchCamera` functions.
-   * @default the first front camera
-   */
-  captureDeviceId: string;
-};
-
 export type ScreencastOptions<MetadataType extends Metadata> = {
   /**
    * Resolution + fps of screencast track, one of: `VGA`, `HD5`, `HD15`, `FHD15`, `FHD30`.
@@ -192,27 +75,6 @@ export type ScreencastOptions<MetadataType extends Metadata> = {
    */
   maxBandwidth: TrackBandwidthLimit;
 };
-
-export type CaptureDevice = {
-  id: string;
-  name: string;
-  isFrontFacing: boolean;
-  isBackFacing: boolean;
-};
-
-export type EndpointsUpdateEvent<
-  MetadataType extends Metadata,
-  VideoTrackMetadataType extends Metadata,
-  AudioTrackMetadataType extends Metadata,
-> = {
-  EndpointsUpdate: Endpoint<
-    MetadataType,
-    VideoTrackMetadataType,
-    AudioTrackMetadataType
-  >[];
-};
-
-export type IsCameraOnEvent = { IsCameraOn: boolean };
 
 export type IsScreencastOnEvent = { IsScreencastOn: boolean };
