+++ packages/react-native-client/src/index.tsx
@@ -1,5 +1,11 @@
+export type {
+  Peer,
+  Track,
+  TrackType,
+  VadStatus,
+  EncodingReason,
+} from './hooks/usePeers';
 export { usePeers } from './hooks/usePeers';
-export type { Peer } from './hooks/usePeers';
 
 export type {
   AudioOutputDevice,
@@ -11,6 +17,11 @@ export { useAudioSettings } from './hooks/useAudioSettings';
 export type { BandwidthEstimationEvent } from './hooks/useBandwidthEstimation';
 export { useBandwidthEstimation } from './hooks/useBandwidthEstimation';
 
+export type {
+  CaptureDevice,
+  VideoQuality,
+  CameraConfig,
+} from './hooks/useCamera';
 export { useCamera } from './hooks/useCamera';
 
 export type {
@@ -36,9 +47,10 @@ export {
 
 export { connect, leaveRoom } from './common/client';
 
-export { VideoPreviewView } from './components/VideoPreviewView';
 export type { VideoPreviewViewProps } from './components/VideoPreviewView';
-export { VideoRendererView } from './components/VideoRendererView';
+export { VideoPreviewView } from './components/VideoPreviewView';
+
 export type { VideoRendererProps } from './components/VideoRendererView';
+export { VideoRendererView } from './components/VideoRendererView';
 
 export * from './RNFishjamClient.types';
